Split the ContractCallServicePrecompileTest into two classes based on the readonly and modification tests
